+++ skore/src/skore/_plugins/hub/artifact/pickle/pickle.py
@@ -47,7 +47,7 @@ class Pickle(Artifact):
         -----
         The report is pickled without its cache, to avoid salting the checksum.
         """
-        reports = [self.report] + getattr(self.report, "estimator_reports_", [])
+        reports = [self.report] + getattr(self.report, "reports_", [])
         reports_with_cache = [
             (report, report._cache) for report in reports if hasattr(report, "_cache")
         ]