+++ skore/src/skore/_sklearn/_cross_validation/inspection_accessor.py
@@ -64,7 +64,7 @@ class _InspectionAccessor(_BaseAccessor[CrossValidationReport], DirNamesMixin):
                     report.inspection.coefficients()
                     .coefficients.copy()
                     .assign(split=split_idx)
-                    for split_idx, report in enumerate(self._parent.estimator_reports_)
+                    for split_idx, report in enumerate(self._parent.reports_)
                 ],
                 ignore_index=True,
             ),
@@ -120,7 +120,7 @@ class _InspectionAccessor(_BaseAccessor[CrossValidationReport], DirNamesMixin):
             Has no effect if the estimator is not a :class:`~sklearn.pipeline.Pipeline`.
 
         metric : str, callable, scorer, or list of such instances or dict of such \
-            instances, default=None
+                instances, default=None
             The metric to pass to :func:`sklearn.inspection.permutation_importance`.
 
             - if ``None``, a suitable default will be used.
@@ -264,7 +264,7 @@ class _InspectionAccessor(_BaseAccessor[CrossValidationReport], DirNamesMixin):
                         n_jobs=n_jobs,
                         seed=seed,
                     ).importances.assign(split=split_idx)
-                    for split_idx, report in enumerate(self._parent.estimator_reports_)
+                    for split_idx, report in enumerate(self._parent.reports_)
                 ],
                 ignore_index=True,
             ),
@@ -316,7 +316,7 @@ class _InspectionAccessor(_BaseAccessor[CrossValidationReport], DirNamesMixin):
                     report.inspection.impurity_decrease()
                     .importances.copy()
                     .assign(split=split_idx)
-                    for split_idx, report in enumerate(self._parent.estimator_reports_)
+                    for split_idx, report in enumerate(self._parent.reports_)
                 ],
                 ignore_index=True,
             ),