+++ skore/tests/unit/project/test_project.py
@@ -121,22 +121,6 @@ class TestProject:
             "workspace": "<workspace>",
         }
 
-    def test_init_local_unknown_plugin(self, monkeypatch, tmp_path):
-        monkeypatch.undo()
-        monkeypatch.setattr(
-            "skore._project.plugin.entry_points", lambda **kwargs: EntryPoints([])
-        )
-
-        with raises(
-            ValueError,
-            match=escape(
-                "The mode `local` is not supported. "
-                "You can install the required dependencies with:\n"
-                '    pip install "skore[local]"'
-            ),
-        ):
-            Project(mode="local", name="<name>")
-
     def test_init_local_missing_optional_dependency(self, monkeypatch):
         fake_library_name = uuid4().hex
 
@@ -207,22 +191,6 @@ class TestProject:
             "name": "<name>",
             "tracking_uri": "<uri>",
         }
-
-    def test_init_hub_unknown_plugin(self, monkeypatch, tmp_path):
-        monkeypatch.undo()
-        monkeypatch.setattr(
-            "skore._project.plugin.entry_points", lambda **kwargs: EntryPoints([])
-        )
-
-        with raises(
-            ValueError,
-            match=escape(
-                "The mode `hub` is not supported. "
-                "You can install the required dependencies with:\n"
-                '    pip install "skore[hub]"'
-            ),
-        ):
-            Project(mode="hub", name="<workspace>/<name>")
 
     def test_init_exception_wrong_ml_task(self, monkeypatch):
         """If the underlying Project implementation contains reports with